requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
beautifulsoup4
lxml
python-dotenv==1.0.1 